requirements.txt
python-docx==1.1.0
pandas==2.1.3
python-dotenv==1.0.0
